Split failover tests into per-class files
Separate RadarVerifiedHostFailoverTests.swift into
RadarFailoverHostSelectorTests.swift and
RadarFailoverAPICoordinatorTests.swift so each file matches the class
under test.
